Update to deploy on netlify
