+++ netlify.toml
@@ -1,0 +1,2 @@
+[build]
+  functions = "functions"
